Fix #2 - correctness issue in dom.
Based on the patch provided by Sebastian Graf.
"The loop has to compare to the current minimum,
not the value it started with (which is the DFS number of the node in question).

Also added a few comments to make the code more readable
and an utility function to output a graph as dot file.